전남대 BE 조홍식 5주차 과제제출 (step3) (#417)
* docs: README 작성

* feat: Kakao인증관련 DTO 및 mapper 클래스 작성

아직 accessToken만 사용하는 단계라, kakaoTokenResponse은 구현만해둔상태입니다.

* Feat: kakao 인가코드 받기

* feat: kakao access Token 받기

* feat: kakao 인가코드주소 받기

* feat: kakao token 요청

* docs: 요구사항 작성

* refac: 메서드 네이밍 변경

* feat: 주문 관련 DTO 및 mapper 구현

* feat: 카카오 메시지관련 예외 클래스 구현

* feat: 카카오토큰 응답 Dto 구현

* feat: order repository 구현

* feat: order service구현

* feat: order entity 구현

* feat: 카카오 메시지 관련 util 클래스 구현

* feat: order controller 구현

* fix: 일단 map을 통한 memberId 별 token 저장

* fix: 클래스 네이밍 변경

* feat: kakaoRepository 구현

* feat: kakaoService 구현

* feat: kakao Entity 구현

* feat: kakao api 관련 예외

* feat: 액세스토큰 map에서 db로 저장함에 따른 메서드 추가

* refac: 인가, 토큰 발급관련 코드 리팩토링

* refac: 토큰 저장방식 변경에 따른 호출함수 변경

* feat: swagger 사용

+++ src/main/java/gift/controller/wish/WishController.java
@@ -5,6 +5,8 @@ import gift.config.LoginUser;
 import gift.controller.auth.AuthController;
 import gift.controller.auth.LoginResponse;
 import gift.service.WishService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,6 +24,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@Tag(name = "Wish", description = "Wish API")
 @RequestMapping("api/wishes")
 public class WishController {
 
@@ -32,6 +35,7 @@ public class WishController {
     }
 
     @GetMapping
+    @Operation(summary = "get All wishes", description = "모든 위시리스트 조회")
     public ResponseEntity<Page<WishResponse>> getAllWishes(
         @RequestParam(defaultValue = "0") int page, @RequestParam(defaultValue = "5") int size) {
         Pageable pageable = PageRequest.of(page, size);
@@ -39,6 +43,7 @@ public class WishController {
     }
 
     @GetMapping("/{memberId}")
+    @Operation(summary = "get All wishes by member", description = "멤버의 모든 위시리스트 조회")
     public ResponseEntity<Page<WishResponse>> getWishes(@PathVariable UUID memberId,
         @RequestParam(defaultValue = "0") int page,
         @RequestParam(defaultValue = "5") int size) {
@@ -48,12 +53,14 @@ public class WishController {
     }
 
     @PostMapping("/{memberId}")
+    @Operation(summary = "create wish", description = "위시리스트 생성")
     public ResponseEntity<WishResponse> createWish(@LoginUser LoginResponse member,
         @PathVariable UUID memberId, @RequestBody WishCreateRequest wish) {
         return ResponseEntity.status(HttpStatus.CREATED).body(wishService.save(memberId, wish));
     }
 
     @PutMapping("/{memberId}/{productId}")
+    @Operation(summary = "modify wish", description = "위시리스트 수정")
     public ResponseEntity<WishResponse> updateWish(@LoginAdmin LoginResponse member,
         @PathVariable UUID memberId, @PathVariable UUID productId,
         @RequestBody WishUpdateRequest wish) {
@@ -62,8 +69,10 @@ public class WishController {
     }
 
     @DeleteMapping("/{memberId}/{productId}")
-    public ResponseEntity<Void> deleteProduct(@LoginAdmin LoginResponse member,
+    @Operation(summary = "delete wish", description = "위시리스트 삭제")
+    public ResponseEntity<Void> deleteProduct(@LoginUser LoginResponse member,
         @PathVariable UUID memberId, @PathVariable UUID productId) {
+        AuthController.validateUserOrAdmin(member, memberId);
         wishService.delete(memberId, productId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
     }
